tor-1132 korjauksia pullarin kommenttien perusteella
